CORE-751: Split out BRCryptoPrivate.h
Create individual BRCryptoXYZPrivate.h headers; move structure definitions from BRCryptoXYZ.c into the corresponding private header.
