+++ crypto/BRCryptoAccount.c
@@ -9,14 +9,7 @@
 //  See the CONTRIBUTORS file at the project root for a list of contributors.
 
 #include <pthread.h>
-
-#include "BRCryptoAccount.h"
-#include "BRCryptoPrivate.h"
-
-#include "support/BRBIP32Sequence.h"
-#include "support/BRBIP39Mnemonic.h"
-#include "support/BRKey.h"
-#include "ethereum/BREthereum.h"
+#include "BRCryptoAccountP.h"
 #include "generic/BRGenericRipple.h"
 
 static pthread_once_t  _accounts_once = PTHREAD_ONCE_INIT;
@@ -38,17 +31,6 @@ randomBytes (void *bytes, size_t bytesCount);
 
 static void
 cryptoAccountRelease (BRCryptoAccount account);
-
-struct BRCryptoAccountRecord {
-    BRMasterPubKey btc;
-    BREthereumAccount eth;
-    BRGenericAccount xrp;
-    // ...
-
-    char *uids;
-    uint64_t timestamp;
-    BRCryptoRef ref;
-};
 
 IMPLEMENT_CRYPTO_GIVE_TAKE (BRCryptoAccount, cryptoAccount);
