+++ crypto/BRCryptoPeer.c
@@ -13,7 +13,6 @@
 #include "support/BRInt.h"
 #include "BRCryptoPeer.h"
 
-
 /** Forward Declarations */
 static void
 cryptoPeerRelease (BRCryptoPeer peer);
