+++ crypto/BRCryptoUnit.c
@@ -9,7 +9,7 @@
 //  See the CONTRIBUTORS file at the project root for a list of contributors.
 
 #include "BRCryptoUnit.h"
-#include "BRCryptoPrivate.h"
+#include "support/BRArray.h"
 
 #include <stdlib.h>
 #include <string.h>
