+++ crypto/BRCryptoWalletManagerP.h
@@ -1,0 +1,130 @@
+//
+//  BRCryptoWalletManagerP.h
+//  BRCrypto
+//
+//  Created by Michael Carrara on 6/19/19.
+//  Copyright © 2019 breadwallet. All rights reserved.
+//
+//  See the LICENSE file at the project root for license information.
+//  See the CONTRIBUTORS file at the project root for a list of contributors.
+
+#ifndef BRCryptoWalletManagerP_h
+#define BRCryptoWalletManagerP_h
+
+#include <pthread.h>
+
+#include "BRCryptoBase.h"
+#include "BRCryptoNetwork.h"
+#include "BRCryptoAccount.h"
+#include "BRCryptoWallet.h"
+#include "BRCryptoWalletManager.h"
+
+#include "ethereum/BREthereum.h"
+#include "bitcoin/BRWalletManager.h"
+#include "generic/BRGenericWalletManager.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct BRCryptoWalletManagerRecord {
+    pthread_mutex_t lock;
+
+    BRCryptoBlockChainType type;
+    union {
+        BRWalletManager btc;
+        BREthereumEWM eth;
+        BRGenericWalletManager gen;
+    } u;
+
+    BRCryptoCWMListener listener;
+    BRCryptoCWMClient client;
+    BRCryptoNetwork network;
+    BRCryptoAccount account;
+    BRCryptoAddressScheme addressScheme;
+
+    BRCryptoWalletManagerState state;
+
+    /// The primary wallet
+    BRCryptoWallet wallet;
+
+    /// All wallets
+    BRArrayOf(BRCryptoWallet) wallets;
+    char *path;
+
+    BRCryptoRef ref;
+};
+
+ /// MARK: - WalletManager
+
+private_extern BRWalletManagerClient
+cryptoWalletManagerClientCreateBTCClient (OwnershipKept BRCryptoWalletManager cwm);
+
+private_extern BREthereumClient
+cryptoWalletManagerClientCreateETHClient (OwnershipKept BRCryptoWalletManager cwm);
+
+private_extern BRGenericClient
+cryptoWalletManagerClientCreateGENClient (OwnershipKept BRCryptoWalletManager cwm);
+
+
+private_extern BRCryptoWalletManagerState
+cryptoWalletManagerStateInit(BRCryptoWalletManagerStateType type);
+
+private_extern BRCryptoWalletManagerState
+cryptoWalletManagerStateDisconnectedInit(BRDisconnectReason reason);
+
+private_extern void
+cryptoWalletManagerSetState (BRCryptoWalletManager cwm,
+                             BRCryptoWalletManagerState state);
+
+
+private_extern void
+cryptoWalletManagerStop (BRCryptoWalletManager cwm);
+
+private_extern BRWalletManager
+cryptoWalletManagerAsBTC (BRCryptoWalletManager manager);
+
+private_extern BREthereumEWM
+cryptoWalletManagerAsETH (BRCryptoWalletManager manager);
+
+private_extern BRCryptoBoolean
+cryptoWalletManagerHasBTC (BRCryptoWalletManager manager,
+                           BRWalletManager bwm);
+
+private_extern BRCryptoBoolean
+cryptoWalletManagerHasETH (BRCryptoWalletManager manager,
+                           BREthereumEWM ewm);
+
+private_extern BRCryptoBoolean
+cryptoWalletManagerHasGEN (BRCryptoWalletManager manager,
+                           BRGenericWalletManager gwm);
+
+private_extern BRCryptoWallet
+cryptoWalletManagerFindWalletAsBTC (BRCryptoWalletManager manager,
+                                    BRWallet *btc);
+
+private_extern BRCryptoWallet
+cryptoWalletManagerFindWalletAsETH (BRCryptoWalletManager manager,
+                                    BREthereumWallet eth);
+
+private_extern BRCryptoWallet
+cryptoWalletManagerFindWalletAsGEN (BRCryptoWalletManager cwm,
+                                    BRGenericWallet gen);
+
+private_extern void
+cryptoWalletManagerAddWallet (BRCryptoWalletManager cwm,
+                              BRCryptoWallet wallet);
+
+private_extern void
+cryptoWalletManagerRemWallet (BRCryptoWalletManager cwm,
+                              BRCryptoWallet wallet);
+
+extern void
+cryptoWalletManagerHandleTransferGEN (BRCryptoWalletManager cwm,
+                                      BRGenericTransfer transferGeneric);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* BRCryptoWalletManagerP_h */
