+++ include/BRCryptoHash.h
@@ -7,7 +7,6 @@
 //
 //  See the LICENSE file at the project root for license information.
 //  See the CONTRIBUTORS file at the project root for a list of contributors.
-//
 
 #ifndef BRCryptoHash_h
 #define BRCryptoHash_h
